requirements.txt
langchain==0.1.0
langchain-openai==0.0.2
langchain-google-genai==0.0.5
langchain-community==0.0.10
pymongo==4.6.1
python-dotenv==1.0.0
tiktoken==0.5.2
numpy==1.24.3
pydantic==2.5.0
azure-cosmos==4.5.1